Preparing 1.2.1

+++ diozero-core/src/main/java/com/diozero/util/MmapBufferNative.java
@@ -4,8 +4,8 @@ package com.diozero.util;
  * #%L
  * Organisation: diozero
  * Project:      Device I/O Zero - Core
- * Filename:     MmapBufferNative.java  
- * 
+ * Filename:     MmapBufferNative.java
+ *
  * This file is part of the diozero project. More information about this project
  * can be found at http://www.diozero.com/
  * %%
@@ -17,10 +17,10 @@ package com.diozero.util;
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -35,7 +35,8 @@ public class MmapBufferNative {
 	static {
 		LibraryLoader.loadSystemUtils();
 	}
-	
+
 	public static native MmapByteBuffer createMmapBuffer(String path, int offset, int length);
-	public static native void closeMmapBuffer(int fd, int mapPtr, int length);
+
+	public static native void closeMmapBuffer(int fd, long mapPtr, int length);
 }
