+++ diozero-core/src/main/java/com/diozero/util/MmapByteBuffer.java
@@ -4,8 +4,8 @@ package com.diozero.util;
  * #%L
  * Organisation: diozero
  * Project:      Device I/O Zero - Core
- * Filename:     MmapByteBuffer.java  
- * 
+ * Filename:     MmapByteBuffer.java
+ *
  * This file is part of the diozero project. More information about this project
  * can be found at http://www.diozero.com/
  * %%
@@ -17,10 +17,10 @@ package com.diozero.util;
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -35,11 +35,11 @@ import java.nio.ByteBuffer;
 
 public class MmapByteBuffer {
 	private int fd;
-	private int address;
+	private long address;
 	private int length;
 	private ByteBuffer buffer;
 
-	public MmapByteBuffer(int fd, int address, int length, ByteBuffer buffer) {
+	public MmapByteBuffer(int fd, long address, int length, ByteBuffer buffer) {
 		this.fd = fd;
 		this.address = address;
 		this.length = length;
@@ -49,15 +49,15 @@ public class MmapByteBuffer {
 	public int getFd() {
 		return fd;
 	}
-	
-	public int getAddress() {
+
+	public long getAddress() {
 		return address;
 	}
-	
+
 	public int getLength() {
 		return length;
 	}
-	
+
 	public ByteBuffer getBuffer() {
 		return buffer;
 	}
